Added a dark backdrop to the bigCards page
Which can be clicked to return to the splash page. Also fixed a bug where hovering over buttons would cause them to flash in and out and another bug where if you re-entered the splash page with your mouse over a card, it wouldn't be tinted and would tint once the mouse left.